feat(grow): exit quietly if not enough free space
Previously, ec2-macos-util's grow command would error out with exit
status 1 if there wasn't enough free space to grow the specified disk.
This wasn't proper behavior since a lack of free space is an expected
case and should be handled as such.

This change adds an extra check in the grow command to see if
GrowContainer funciton returned a FreeSpaceError and exit quietly if it
did. This will cause ec2-macos-utils to log an extra message and return
with exit status 0.